Refactored code to increase maintainability
